+++ frontend/src/components/muiTheme.js
@@ -28,6 +28,17 @@ export function createMuiTheme(primaryColor) {
       MuiTypography: {
         fontFamily: 'Roboto, Helvetica, Arial, sans-serif',
       },
+      MuiDialog: {
+        styleOverrides: {
+          paper: {
+            color: 'var(--div-text-color)',
+            backgroundColor: 'var(--div-bg-color)',
+            border: '2px solid var(--div-border-color)',
+            borderRadius: 'var(--div-border-radius)',
+            boxShadow: '2px 2px 5px var(--div-shadow-color)',
+          },
+        },
+      },
       MuiTooltip: {
         defaultProps: {
           placement: 'top-start',